feat: adicionar autentição de token para api

+++ src/Controller/UserController.php
@@ -2,9 +2,9 @@
 
 namespace App\Controller;
 
-use App\Entity\User;
 use App\Service\UserService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Attribute\Route;
@@ -14,29 +14,47 @@ class UserController extends AbstractController
 
     private UserService $userService;
 
-    public function __construct(UserService $userService)
-    {
+    public function __construct(
+        UserService $userService,
+
+        #[Autowire('%env(API_TOKEN)%')]
+        private string $apiToken
+
+    ) {
         $this->userService = $userService;
     }
 
     //Listar todos os usuários
     #[Route('/users', methods: ['GET'],  name: 'user.index')]
-    public function index(): JsonResponse
+    public function index(Request $request): JsonResponse
     {
-        return $this->json($this->userService->getAllUsers());
+        if ($request->headers->get('Authorization') !== "Bearer {$this->apiToken}") {
+            return $this->json(['msg' => 'Acesso negado: token invalido'], 401);
+        }
+
+        return $this->json($this->userService->getAllUsers(), 200, [], ['groups' => 'user_read']);
     }
 
     // Criar Usuários
     #[Route('/users', methods: ['POST'], name: 'usersCreate.create')]
     public function create(Request $request): JsonResponse
     {
+        if ($request->headers->get('Authorization') !== "Bearer {$this->apiToken}") {
+            return $this->json(['msg' => "Acesso negado: token inválido"], 401);
+        }
+
         $data = $request->toArray();
 
-        $this->userService->createUser($data['nome'], $data['email'], $data['senha']);
+        $user = $this->userService->createUser($data['nome'], $data['email'], $data['senha']);
 
         return $this->json(
             [
-                'msg' => 'Usuário criado com sucesso!'
+                'msg' => 'Usuário criado com sucesso!',
+                'user' => [
+                    'id' => $user->getId(),
+                    'nome' => $user->getNome(),
+                    'email' => $user->getEmail(),
+                ]
             ],
             201
         );
@@ -47,24 +65,39 @@ class UserController extends AbstractController
     #[Route('/users/{id}', methods: ['PUT'], name: 'usersUpdate.update')]
     public function update(int $id, Request $request): JsonResponse
     {
+        if ($request->headers->get('Authorization') !== "Bearer {$this->apiToken}") {
+            return $this->json(['msg' => 'Acesso negado: token inválidio'], 401);
+        }
+
         $data = $request->toArray();
 
-        $users = $this->userService->updateUser($id, $data['nome'] ?? null, $data['email'] ?? null, $data['senha'] ?? null);
+        $user = $this->userService->updateUser($id, $data['nome'] ?? null, $data['email'] ?? null, $data['senha'] ?? null);
 
-        if (!$users) {
+        if (!$user) {
             return $this->json([
                 'msg' => "Usuário não encontrado"
             ], 404);
         }
 
         return $this->json([
-            'msg' => "Usuário com email {$users->getEmail()} foi editado com sucesso"
+            'msg' => "Usuário com email {$user->getEmail()} foi editado com sucesso",
+            'user' => [
+                'id' => $user->getId(),
+                'nome' => $user->getNome(),
+                'email' => $user->getEmail(),
+            ]
         ], 200);
     }
 
+    //Deletar Usuários
     #[Route('/users/{id}', methods: ['DELETE'], name: 'usersDelete.delete')]
-    public function delete(int $id): JsonResponse
+    public function delete(int $id, Request $request): JsonResponse
     {
+
+        if ($request->headers->get('Authorization') !== "Bearer {$this->apiToken}") {
+            return $this->json(['msg' => "Acesso negado: token inválido"], 401);
+        }
+
         $deleted = $this->userService->deleteUser($id);
 
         if (!$deleted) {
